Creates utils.js
