+++ src/utils.js
@@ -1,0 +1,15 @@
+const getRandomArrayItem = (array) => {
+  const randomIndex = getRandomIntegerNumber(0, array.length);
+
+  return array[randomIndex];
+};
+
+const getRandomIntegerNumber = (min, max) => {
+  return min + Math.floor(max * Math.random());
+};
+
+const getRandomDecimal = (min, max) => {
+  return ((max - min) * Math.random()).toFixed(1);
+};
+
+export {getRandomArrayItem, getRandomIntegerNumber, getRandomDecimal};
